LayoutTests/media/media-source/media-source-first-sample-pts-non-zero-canplay-without-renderer-expected.txt

This test verifies that the "canplay" event fires on a video element with a MediaSource that loads a video track whose first sample has a presentation timestamp greater than 0 when the video element is not in the DOM.
EVENT(sourceopen)
END OF TEST
